+++ IcVibracoes/Controllers/FiniteElementController.cs
@@ -30,9 +30,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("rectangular")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(
@@ -58,9 +60,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("rectangular/dynamic-vibration-absorber")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(
@@ -85,9 +89,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("rectangular/piezoelectric")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(
@@ -112,9 +118,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("circular")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(
@@ -139,9 +147,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("circular/dynamic-vibration-absorber")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(
@@ -166,9 +176,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("circular/piezoelectric")]
         public async Task<ActionResult<FiniteElementResponse>> CalculateVibration(