+++ IcVibracoes/Controllers/RigidBodyController.cs
@@ -23,9 +23,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("one-degree-freedom")]
         public async Task<ActionResult<OneDegreeOfFreedomResponse>> CalculateVibration(
@@ -50,9 +52,11 @@ namespace IcVibracoes.Controllers
         /// <returns>A file with analysis result.</returns>
         /// <response code="201">Returns the newly created files.</response>
         /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
         /// <response code="501">If some resource is not implemented.</response>
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
         [HttpPost("two-degree-freedom")]
         public async Task<ActionResult<TwoDegreesOfFreedomResponse>> CalculateVibration(